requirements.txt
google-api-python-client==2.111.0
google-generativeai==0.3.2
instaloader==4.10.3
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.31.0
pandas==2.1.4
